--help

+++ OffersPlatform.API/Controllers/v1/Users/UserProfileController.cs
@@ -1,13 +1,16 @@
+using System.Security.Claims;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using OffersPlatform.Application.DTOs;
 using OffersPlatform.Application.Features.Admin.Users.Commands.DeleteUser;
 using OffersPlatform.Application.Features.Admin.Users.Queries.GetUserById;
+using OffersPlatform.Application.Features.Users.Profile.Commands.UpdateUserBalance;
 
 namespace OffersPlatform.API.Controllers.v1.Users;
 
 [ApiController]
+[Authorize]
 [Route("api/v1/users/profile")]
 public class UserProfileController : ControllerBase
 {
@@ -18,28 +21,42 @@ public class UserProfileController : ControllerBase
         _mediator = mediator;
     }
 
+    private Guid GetUserId()
+    {
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (userId is null)
+        {
+            throw new Exception("Invalid user id");
+        }
+        return Guid.Parse(userId);
+    }
     
     [HttpGet("")]
-    public async Task<IActionResult> GetUser(Guid id, CancellationToken cancellationToken)
+    public async Task<IActionResult> GetUserProfile(Guid id, CancellationToken cancellationToken)
     {
         var query = new GetUserByIdQuery(id);
         var user = await _mediator.Send(query, cancellationToken);
         return Ok(user);
     }
-    
-    [Authorize (Roles = "Admin, Customer")]
-    [HttpPut("")]
-    public async Task<IActionResult> UpdateProfile(Guid id, [FromBody] UserUpdateDto userUpdateDto)
+
+    [HttpPut("balance")]
+    public async Task<IActionResult> AddToBalance(decimal balance, CancellationToken cancellationToken = default)
     {
-        return Ok();
+        var command = new UpdateUserBalanceCommand(GetUserId(), balance);
+        var result = await _mediator.Send(command, cancellationToken);
+        return Ok(result);
     }
-    
-    [Authorize (Roles = "Admin, Customer")]
+
     [HttpDelete("")]
     public async Task<IActionResult> DeleteProfile(Guid id, CancellationToken cancellationToken)
     {
+        if (id == GetUserId())
+        {
+            return BadRequest("Invalid user id");
+        }
         var query = new DeleteUserCommand(id);
         await _mediator.Send(query, cancellationToken);
         return Ok();
     }
+    
 }
